requirements.txt
langchain
langchain-openai
langchain-community
langgraph
chromadb
duckduckgo-search
openai
tiktoken
python-dotenv